[REFACTOR] Post, Category 리팩토링

+++ src/main/java/com/cotato/kampus/domain/comment/dto/CommentDto.java
@@ -17,7 +17,7 @@ public record CommentDto (
 	CommentStatus commentStatus,
 	Anonymity anonymity,
 	Long reports,
-	Long anonymousNumber,
+	Integer anonymousNumber,
 	Long parentId,
 	Long targetId,
 	LocalDateTime createdTime
